added smarts

+++ syncPreAlpha.py
@@ -6,23 +6,24 @@ import requests
 from dotenv import load_dotenv
 
 
-def _isValidDir(directory):
+def _validDir(directory):
     dirname = os.path.dirname(directory)
-    if os.path.exists(dirname): return True
-    print(
-        "Not a valid directory name, make sure you're not using ~ or any shorthand"
-    )
-    return False
+    if not os.path.exists(dirname):
+        print(
+            "Not a valid directory name, make sure you're not using ~ or any shorthand"
+        )
+        return 1
+    return 0
 
 
-def _isHubInstalled():
+def _hubInstalled():
 
     # check for hub
     print("Checking for hub...")
     if os.system("which hub"):
         print("Need to install hub to create PRs. Run `brew install hub`.")
-        return False
-    return True
+        return 1
+    return 0
 
 
 def _cloneRock(rockDir):
@@ -35,15 +36,20 @@ def _cloneRock(rockDir):
     return 1
 
 
-def _cleanup(deleteRepo):
+def _checkout(branch):
 
-    # delete the Rock repo
-    if deleteRepo:
-        os.system("rm -rf Rock")
+    # checkout and update branch
+    r = os.popen("git checkout {}".format(branch)).read()
+    if "up to date" not in r: os.system("git pull")
+
+
+def _cleanup(deleteRepo, deleteRemote):
+
+    # delete the local Rock repo
+    if deleteRepo: os.system("rm -rf Rock")
 
     # delete the local and remote branches
-    os.system("git push --delete origin sync-pre-alpha")
-    os.system("git checkout alpha")
+    if deleteRemote: os.system("git push --delete origin sync-pre-alpha")
     os.system("git branch -D sync-pre-alpha")
 
 
@@ -51,18 +57,22 @@ def _createPreAlpha(rockDir):
 
     # switch to Rock directory
     os.chdir(rockDir)
-    # checkout alpha branch
-    os.system("git checkout alpha")
-    os.system("git pull")
-    # set up Spark as remote
-    os.system(
-        "git remote add SparkDevNetwork https://github.com/SparkDevNetwork/Rock.git"
-    )
+
     # get Spark remote branches
+    if "SparkDevNetwork" not in os.popen("git remote").read():
+        # set up Spark as remote
+        os.system(
+            "git remote add SparkDevNetwork https://github.com/SparkDevNetwork/Rock.git"
+        )
     os.system("git fetch SparkDevNetwork")
+
     # create pre-alpha branch from spark off of NewSpring alpha
-    os.system(
-        "git checkout -b sync-pre-alpha SparkDevNetwork/pre-alpha-release")
+    if "sync-pre-alpha" not in os.popen("git branch").read():
+        os.system(
+            "git checkout -b sync-pre-alpha SparkDevNetwork/pre-alpha-release")
+
+    # push pre alpha branch
+    os.system("git push --set-upstream origin sync-pre-alpha")
 
 
 def _getSHA():
@@ -71,20 +81,28 @@ def _getSHA():
     return os.popen("git rev-parse --short HEAD").read().strip("\n")
 
 
+def _getBranch():
+
+    # this will get the current branch name
+    return os.popen("git rev-parse --abbrev-ref HEAD").read().strip("\n")
+
+
 def _merge(destination, source):
 
-    # checkout destination branch
-    os.system("git checkout {}".format(destination))
-    os.system("git pull")
-    commit = _getSHA()
-    # merge alpha into pre-alpha branch
+    _checkout(source)
+    _checkout(destination)
+
+    # merge source into destination branch
     os.system("git merge --no-ff {} -m \"Merge from NewSpring/{}\"".format(
         source, source))
-    # if the commit hash didn't change, that means the merge failed
-    if commit == _getSHA():
-        print("Merge failed. Check for conflicts")
+    # bail out if merge failed
+    if os.system("git merge --stat > /dev/null 2>&1"):
+        deletedBySource = os.popen(
+            "git diff --name-only --diff-filter=UD").read().split("\n")
+        print(deletedBySource)
+        print("Fix conflicts and run again.")
         return 1
-    # push pre alpha branch
+    # push destination branch
     os.system("git push --set-upstream origin {}".format(destination))
     return 0
 
@@ -111,10 +129,10 @@ def _buildCheck(branch):
     while commit not in buildCommit:
         elapsed = round(time.perf_counter()) - start
         if elapsed > 30:
-            print("Timeout. {} not building.".format(branch))
+            print("\nTimeout. Branch '{}' not building.".format(branch))
             return 1
         print(
-            "\rVerifying correct AppVeyor build ({}s)...".format(elapsed),
+            "\rVerifying automatic AppVeyor build ({}s)...".format(elapsed),
             end="")
         r = requests.get(
             "https://ci.appveyor.com/api/projects/NewSpring/rock/branch/{}".
@@ -129,7 +147,7 @@ def _buildStatus(branch):
     # wait on Appveyor build to pass or fail
     status = ""
     start = round(time.perf_counter())
-    while status not in ["success", "fail"]:
+    while status not in ["success", "failed"]:
         elapsed = round(time.perf_counter()) - start
         print("\rWaiting on AppVeyor build ({}s)...".format(elapsed), end="")
         r = requests.get(
@@ -166,60 +184,66 @@ def _deploy(branch, envID, authKey):
 def _safeMerge(destination, source):
 
     # first merge destination branch into source so we can test the build
-    _merge(source, destination)
+    if _merge(source, destination): return 1
 
     # run appveyor build on source branch if it's not automatic
     if _buildCheck(source): _build(source, os.getenv("APPVEYOR_KEY"))
 
     # after build passes, we can safely merge
-    if not _buildStatus(source):
+    if _buildStatus(source): return 1
 
-        # merge source into destination
-        _merge(destination, source)
-
-
-def _pr(base, head):
-
-    os.system(
-        "hub pull-request -b {} -h {} -m \"Merge from NewSpring/{}\"".format(
-            base, head, head))
-    print("Created {} -> {} PR. You will have to merge and deploy manually.".
-          format(head, base))
+    # merge source into destination
+    _merge(destination, source)
+    return 0
 
 
-def sync(rockDir="Rock", safe=False):
+def _pr(base):
+
+    head = _getBranch()
+    if os.system(
+            "hub pull-request -b {} -m \"Merge from NewSpring/{}\"".format(
+                base, head)):
+        return 1
+    print("Created {} -> {} PR. Ready to merge and deploy {} manually.".format(
+        head, base, base))
+    return 0
+
+
+def sync(rockDir="Rock", safeAlpha=False, safeBeta=False):
     """This will sync Rock pre-alpha with our alpha branch."""
 
     # load environment variables
     load_dotenv()
 
     # first make sure a valid directory was passed
-    if not _isValidDir(rockDir): return 1
+    if _validDir(rockDir): return 1
 
     # check that hub is installed
-    if not _isHubInstalled(): return 1
+    if _hubInstalled(): return 1
 
     # if Rock doesn't exist, clone it and delete it later
     deleteRepo = _cloneRock(rockDir)
+    deleteRemote = not safeAlpha
 
     # checkout alpha branch
     _createPreAlpha(rockDir)
 
-    # merge pre alpha into alpha after it builds successfully
-    _safeMerge("alpha", "sync-pre-alpha")
-
-    if safe:
-        # create a PR instead of automatic deploy
-        _pr("beta", "alpha")
+    if safeAlpha:
+        if _pr("alphaTest"): deleteRemote = True
     else:
-        # merge alpha into beta
-        _safeMerge("beta", "alpha")
+        # merge pre alpha into alpha after it builds successfully
+        if _safeMerge("alphaTest", "sync-pre-alpha"): return 1
 
-        _deploy(destination, os.getenv("APPVEYOR_ENV"),
-                os.getenv("APPVEYOR_KEY"))
+        if safeBeta:
+            _pr("beta")
+        else:
+            if _safeMerge("beta", "alphaTest"): return 1
+            _deploy(destination, os.getenv("APPVEYOR_ENV"),
+                    os.getenv("APPVEYOR_KEY"))
 
     # cleanup repo and stale branches
-    _cleanup(deleteRepo)
+    _checkout("alphaTest")
+    _cleanup(deleteRepo, deleteRemote)
 
 
-sync("/Users/michael.neeley/Documents/Projects/Rock", True)
+sync("/Users/michael.neeley/Documents/Projects/Rock", False, True)
